got endpoints changed. still working on it
