+++ polaris/polaris/sep31/urls.py
@@ -1,9 +1,8 @@
 from django.urls import path
-from polaris.sep31 import info, send, transaction, update
+from polaris.sep31 import info, transaction
 
 urlpatterns = [
     path("info", info.info),
-    path("send", send.send),
-    path("transaction", transaction.transaction),
-    path("update", update.update),
+    path("transaction/<transaction_id>", transaction.TransactionAPIView.as_view()),
+    path("transaction", transaction.TransactionAPIView.as_view()),
 ]
